// src_bak/utils/types.ts
export interface ISpell {
  id: string;
  name: string;
  strength: number;
  element: IElement;
  description: string;
}

export type IElement = {
  code: string;
  color: colorType;
  description: string;
  school: ISchool;
};

export type ISchool = {
  id: schoolType;
  description: string;
  name: string;
};

export type IPlayerSpell = ISpell & {
  copy: number;
  selected: boolean;
  owner: "hero" | "enemy";
  updates: IPlayerSpellUpdate[];
  created_at: Date;
};
export type IPlayerResource = IResource & {
  quantity: number;
  created_at: Date;
};
export type INPC = ICharacter & {
  image: string;
  dialogue: string | null;
  created_at: Date;
};
export type IHero = ICharacter & {
  element: IElement;
};
export type IPlayerHero = IHero & { selected: boolean; created_at: Date };

export type IEnemy = ICharacter & {
  element: IElement;
};

export type IEnemyFight = IEnemy & { spells: IPlayerSpell[] };

export type ISpellUpdate = {
  id: string;
  school: ISchool;
  mana: number;
  effect: spellEffectType;
  action: IUpdateAction;
  price: { action: spellUpdatePrice; strength: number | string } | null;
  name: string;
  description: string;
  resource_base: IResource[];
};

export type IPlayerSpellUpdate = ISpellUpdate & {
  created_at: Date;
  resource_base: IResource[];
};

export type IUpdateAction = {
  item: spellUpdateEffect;
  data: number | string;
};

export type PlayerData = {
  id: string;
  experience: number;
  life: number;
  maxlife: number;
  mana: number;
  maxmana: number;
  created_at: Date;
  rank: number;
};
export type Player = {
  data: PlayerData;
  npcs: INPC[];
  heroes: IPlayerHero[];
  spells: IPlayerSpell[];
  updates: IPlayerSpellUpdate[];
  adventures: IPlayerAdventure[];
  resources: IPlayerResource[];
};
// ...
export type ICharacter = {
  id: string;
  name: string;
  description: string;
};

export type IAdventure = {
  id: adventureType;
  type: string;
  name: string;
  image: string;
  stories: IStory;
};
export type IPlayerAdventure = IAdventure & { open: boolean; created_at: Date };
// ...
export type IFight = {
  id: string;
  name: string;
  hero_num: number;
  enemy: IEnemy;
  hero_elements: colorType[];
};

export type FightState = {
  hero: {
    life: number;
    maxLife: number;
    mana: number;
    maxMana: number;
  };
  heroes: IPlayerHero[];
  enemy: IEnemyFight;
  heroDeck: IPlayerSpell[];
  heroCardIndex: number | null;
  heroDrop: IPlayerSpell[];
  heroHand: IPlayerSpell[];
  enemyDeck: IPlayerSpell[];
  enemyDrop: IPlayerSpell[];
  enemyCardIndex: number | null;
  element: colorType;
  elements: colorType[];
};

export type IResource = {
  id: string;
  name: string;
  school: ISchool;
  description: string;
  commonality: number;
};

export type storyType = "dialogue" | "fight" | "reel";

export type adventureType =
  | "character"
  | "story"
  | "arena"
  | "torunament"
  | "event";
export type screenState = "start" | "opening" | "main" | "dialogue";
export type enemyExpLevel =
  | "novice"
  | "apprentice"
  | "practitioner"
  | "master"
  | "grandmaster";
export type colorType =
  | "violet"
  | "grey"
  | "red"
  | "indigo"
  | "green"
  | "blue"
  | "white"
  | "orange"
  | "yellow"
  | "rose"
  | "teal"
  | "sand"
  | "lime"
  | "black"
  | "brown";